Fixed some issues

+++ ROS/src/locmap/locmap_fastslam/include/dbscan.hpp
@@ -4,7 +4,7 @@
 #include <Eigen/Dense>
 
 auto dbscanVectorXf(const std::vector<VectorXf>& data, float eps, int min_pts) 
-        -> std::vector<std::vector<size_t>>;
+        -> std::vector<std::vector<unsigned int>>;
 
 auto dbscan(const std::vector<std::pair<float, float>>& data, float eps, int min_pts) 
-        -> std::vector<std::vector<size_t>>;
+        -> std::vector<std::vector<unsigned int>>;
